# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Majlis Atfal-ul-Ahmadiyya Ghana Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyticsSummary {
  totalMembers: number;
  atfalSughirCount: number;
  atfalKabirCount: number;
  khuddamCount: number;
  newThisMonth: number;
  graduationsThisMonth: number;
}