# lib/api-zod/src/generated/types/circuitSuggestion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Majlis Atfal-ul-Ahmadiyya Ghana Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CircuitSuggestion {
  id: number;
  name: string;
  /** @nullable */
  zone?: string | null;
  usageCount?: number;
}