# lib/api-zod/src/generated/types/graduationRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Majlis Atfal-ul-Ahmadiyya Ghana Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface GraduationRecord {
  id: number;
  memberId: number;
  memberName: string;
  zone: string;
  region: string;
  sector?: string;
  graduatedAt: Date;
  previousWing: string;
  newWing: string;
}