# lib/api-zod/src/generated/types/memberHistory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Majlis Atfal-ul-Ahmadiyya Ghana Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface MemberHistory {
  id: number;
  memberId: number;
  eventType: string;
  description: string;
  /** @nullable */
  previousValue?: string | null;
  /** @nullable */
  newValue?: string | null;
  createdAt: Date;
}